requirements.txt
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
python-decouple==3.6
requests==2.27.1
soupsieve==2.3.2.post1
urllib3==1.26.9
